requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
Django==4.2.1
django-bootstrap5==23.1
django-environ==0.10.0
django-selenium-login==2.0.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
mysqlclient==2.1.1
outcome==1.2.0
PySocks==1.7.1
python-dateutil==2.8.2
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
wsproto==1.2.0
